# Conflicts:
#	CatDogShelter/build.gradle
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/common/entity/QuestionCategoryEntity.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/common/entity/ShelterHeadEntity.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/common/entity/SidoEntity.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/common/entity/SigunguEntity.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/common/util/ReportCategory.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/adoption/command/application/controller/AdoptionPostCommandController.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/adoption/command/application/dto/AdoptionPostCommandDTO.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/adoption/command/application/service/AdoptionPostCommandService.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/adoption/command/domain/aggregate/entity/AdoptionPost/AdoptionPostEntity.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/adoption/command/domain/aggregate/entity/UserEntity.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/adoption/command/domain/repository/AdoptionRepository.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/adoption/query/controller/AdoptionPostQueryController.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/adoption/query/mapper/AdoptionMapper.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/adoption/query/service/AdoptionPostQueryService.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/adoption/query/service/AdoptionPostQueryServiceImpl.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/post/command/application/controller/PostExController.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/post/command/application/service/PostService.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/post/command/domain/aggregate/entity/PostCommentEntity.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/post/command/domain/aggregate/entity/PostFilesEntity.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/post/command/domain/repository/PostCommentRepository.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/post/command/domain/repository/PostFilesRepository.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/post/command/domain/repository/PostRepository.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/user/command/domain/repository/UserJpaRepositoryInterface.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/user/command/domain/repository/UserRepository.java
#	CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/user/query/service/QueryService.java
#	CatDogShelter/src/main/resources/application.yml
#	CatDogShelter/src/test/java/com/backoven/catdogshelter/domain/volunteer/query/service/VolunteerQueryServiceTest.java
#	DB/DDL/CatDogShelter_DDL.sql

+++ CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/donation/command/application/dto/CreateDonationCommentRequest.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CreateDonationCommentRequest {
-    private Long postId;
+    private Integer postId;
     private String content;
-    private Long userId;
+    private Integer userId;
 }
