+++ CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/donation/command/application/service/DonationPostReportCommentCommandService.java
@@ -1,12 +1,12 @@
 package com.backoven.catdogshelter.domain.donation.command.application.service;
 
+import com.backoven.catdogshelter.common.entity.UserEntity;
 import com.backoven.catdogshelter.common.util.DateTimeUtil;
 import com.backoven.catdogshelter.common.util.ReportCategory;
 import com.backoven.catdogshelter.domain.donation.command.domain.aggregate.entity.DonationPostComment;
 import com.backoven.catdogshelter.domain.donation.command.domain.aggregate.entity.DonationPostCommentReport;
 import com.backoven.catdogshelter.domain.donation.command.domain.repository.DonationPostCommentReportRepository;
 import com.backoven.catdogshelter.domain.donation.command.domain.repository.DonationPostCommentRepository;
-import com.backoven.catdogshelter.domain.user.command.domain.aggregate.entity.UserEntity;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,7 +19,7 @@ public class DonationPostReportCommentCommandService {
     private final DonationPostCommentRepository donationPostCommentRepository;
     private final DonationPostCommentReportRepository donationPostCommentReportRepository;
 
-    public void createReportDonationPostComment(Long commentId, ReportCategory category, String detail, UserEntity reporter) {
+    public void createReportDonationPostComment(Integer commentId, ReportCategory category, String detail, UserEntity reporter) {
         DonationPostComment comment = donationPostCommentRepository.findById(commentId)
                 .orElseThrow(() -> new IllegalArgumentException("Comment not found"));
