+++ CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/donation/command/domain/repository/DonationPostCommentRepository.java
@@ -3,5 +3,5 @@ package com.backoven.catdogshelter.domain.donation.command.domain.repository;
 import com.backoven.catdogshelter.domain.donation.command.domain.aggregate.entity.DonationPostComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DonationPostCommentRepository extends JpaRepository<DonationPostComment, Long> {}
+public interface DonationPostCommentRepository extends JpaRepository<DonationPostComment, Integer> {}
