+++ CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/donation/command/domain/repository/DonationPostFilesRepository.java
@@ -3,4 +3,4 @@ package com.backoven.catdogshelter.domain.donation.command.domain.repository;
 import com.backoven.catdogshelter.domain.donation.command.domain.aggregate.entity.DonationPostFiles;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DonationPostFilesRepository extends JpaRepository<DonationPostFiles, Long> {}
+public interface DonationPostFilesRepository extends JpaRepository<DonationPostFiles, Integer> {}
