+++ CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/donation/query/controller/DonationPostQueryController.java
@@ -49,7 +49,7 @@ public class DonationPostQueryController {
 
     //REQ-028 조회 수 증가
     @PostMapping("/{id}/view")
-    public void increaseView(@PathVariable Long id) {
+    public void increaseView(@PathVariable Integer id) {
         service.increaseView(id);
     }
