+++ CatDogShelter/src/main/java/com/backoven/catdogshelter/domain/donation/query/mapper/DonationPostQueryMapper.java
@@ -23,7 +23,7 @@ public interface DonationPostQueryMapper {
     List<DonationPostQueryDTO> selectAllDonationPostsLatest(@Param("limit") int limit);
 
     // REQ-028 조회수 증가
-    void increaseView(@Param("id") Long id);
+    void increaseView(@Param("id") Integer id);
 
     List<DonationPostQueryDTO> selectAllDonationPosts();
